# app/openapi.yaml
openapi: 3.0.0
info:
  title: 'OpenApi Documentation For Portal Vagas'
  description: 'Documentation intended to guide the use of the API to access Portal Vagas data'
  contact:
    email: portal_vagas_admin@gmail.com
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8190'
    description: 'Demo API Server'
paths:
  /api/applications:
    get:
      tags:
        - Applications
      summary: 'Get list of applications'
      description: 'Returns list of applications'
      operationId: getApplicationsList
      parameters:
        -
          $ref: '#/components/parameters/page'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResourceCollection'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
    post:
      tags:
        - Applications
      summary: 'Store new Application'
      description: 'Returns Application data'
      operationId: storeApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResource'
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
  '/api/applications/{id}':
    get:
      tags:
        - Applications
      summary: 'Get Application information'
      description: 'Returns Application data'
      operationId: getApplicationById
      parameters:
        -
          name: id
          in: path
          description: 'Application id'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResource'
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
    put:
      tags:
        - Applications
      summary: 'Update existing Application'
      description: 'Returns updated Application data'
      operationId: updateApplication
      parameters:
        -
          name: id
          in: path
          description: 'Application id'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApplicationRequest'
      responses:
        '202':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResource'
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: 'Resource Not Found'
    delete:
      tags:
        - Applications
      summary: 'Delete existing Application'
      description: 'Deletes a record and returns no content'
      operationId: deleteApplication
      parameters:
        -
          name: id
          in: path
          description: 'Application id'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Successful operation'
          content:
            application/json:
              schema: {  }
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: 'Resource Not Found'
  /api/candidates:
    get:
      tags:
        - Candidates
      summary: 'Get list of candidates'
      description: 'Returns list of candidates'
      operationId: getCandidatesList
      parameters:
        -
          $ref: '#/components/parameters/page'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResourceCollection'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
    post:
      tags:
        - Candidates
      summary: 'Store new candidate'
      description: 'Returns candidate data'
      operationId: storeCandidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCandidateRequest'
      responses:
        '201':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResource'
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
  '/api/candidates/{id}':
    get:
      tags:
        - Candidates
      summary: 'Get candidate information'
      description: 'Returns candidate data'
      operationId: getCandidateById
      parameters:
        -
          name: id
          in: path
          description: 'Candidate id'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResource'
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
    put:
      tags:
        - Candidates
      summary: 'Update existing candidate'
      description: 'Returns updated candidate data'
      operationId: updateCandidate
      parameters:
        -
          name: id
          in: path
          description: 'Candidate id'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCandidateRequest'
      responses:
        '202':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResource'
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: 'Resource Not Found'
    delete:
      tags:
        - Candidates
      summary: 'Delete existing candidate'
      description: 'Deletes a record and returns no content'
      operationId: deleteCandidate
      parameters:
        -
          name: id
          in: path
          description: 'Candidate id'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Successful operation'
          content:
            application/json:
              schema: {  }
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: 'Resource Not Found'
...
components:
  schemas:
...
    CreateApplicationRequest:
      properties:
        candidate_id:
          type: string
          example: '1'
        opportunity_id:
          type: string
          example: '1'
      type: object
    CreateCandidateRequest:
      properties:
        name:
          type: string
          example: name
        email:
          type: string
          example: email@teste.com
        phone:
          type: string
          example: phone
        level:
          type: string
          example: level
        area:
          type: string
          example: area
        expected_salary:
          type: string
          example: '3200'
      type: object
...
    UpdateApplicationRequest:
      properties:
        candidate_id:
          type: string
          example: '1'
        opportunity_id:
          type: string
          example: '1'
      type: object
    UpdateCandidateRequest:
      properties:
        name:
          type: string
          example: name
        email:
          type: string
          example: email@teste.com
        phone:
          type: string
          example: phone
        level:
          type: string
          example: level
        area:
          type: string
          example: area
        expected_salary:
          type: string
          example: expected_salary
      type: object
...
    ApplicationResource:
      properties:
        id:
          type: integer
          example: '1'
        candidate:
          type: string
          example: 'Sample Name'
        opportunity:
          type: string
          example: 'Sample Description'
      type: object
    ApplicationResourceCollection:
      type: array
      items:
        $ref: '#components/schemas/ApplicationResource'
    CandidateResource:
      properties:
        id:
          type: integer
          example: '1'
        name:
          type: string
          example: name
        email:
          type: string
          example: email@teste.com
        phone:
          type: string
          example: phone
        level:
          type: string
          example: level
        area:
          type: string
          example: area
        expected_salary:
          type: integer
          example: '10299'
      type: object
    CandidateResourceCollection:
      type: array
      items:
        $ref: '#components/schemas/ApplicationResource'
...
  parameters:
    page:
      name: page
      in: query
      description: 'A pagina alvo'
      required: true
      schema:
        type: integer
        example: '1'
...